Update Thrift server to use Thrift arg scheme
